pyproject.toml
[project]
name = "sentimentanalysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.3.5",
    "emoji==0.6.0",
    "google-genai>=1.52.0",
    "hf-xet>=1.2.0",
    "langchain>=1.0.8",
    "langchain-core>=1.1.0",
    "langchain-google-genai>=3.1.0",
    "langchain-google-vertexai>=3.0.3",
    "langchain-mistralai>=1.0.1",
    "langsmith>=0.4.46",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "poethepoet>=0.37.0",
    "pydantic>=2.12.4",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.6",
    "streamlit>=1.51.0",
    "tavily>=1.1.0",
    "torch>=2.0.0",
    "transformers>=4.57.1",
]

[tool.poe.tasks]
# Linting tasks
lint = "ruff check ."
format = "ruff format ."
lint-fix = "ruff check --fix ."

_start = "streamlit run app.py"
_dev = "python main.py"
_worker = "uvicorn api:app --reload"

start = ["lint", "_start"]
dev = ["lint", "_dev"]
worker = ["lint", "_worker"]
